implemented linux directory force removal and implemented tests for it

+++ src/utilities/directorycreator.cpp
@@ -6,6 +6,70 @@
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <unistd.h>
+#include <dirent.h>
+#include <cstring>
+
+int remove_directory_force(const char *path)
+{
+    DIR *d = opendir(path);
+    size_t path_len = strlen(path);
+    int r = -1;
+
+    if (d)
+    {
+        struct dirent *p;
+
+        r = 0;
+
+        while (!r && (p=readdir(d)))
+        {
+            int r2 = -1;
+            char *buf;
+            size_t len;
+
+            /* Skip the names "." and ".." as we don't want to recurse on them. */
+            if (!strcmp(p->d_name, ".") || !strcmp(p->d_name, ".."))
+            {
+                continue;
+            }
+
+            len = path_len + strlen(p->d_name) + 2; 
+            buf = (char*) malloc(len);
+
+            if (buf)
+            {
+                struct stat statbuf;
+
+                snprintf(buf, len, "%s/%s", path, p->d_name);
+
+                if (!stat(buf, &statbuf))
+                {
+                    if (S_ISDIR(statbuf.st_mode))
+                    {
+                        r2 = remove_directory_force(buf);
+                    }
+                    else
+                    {
+                        r2 = unlink(buf);
+                    }
+                }
+
+                free(buf);
+            }
+
+            r = r2;
+        }
+
+        closedir(d);
+    }
+
+    if (!r)
+    {
+        r = rmdir(path);
+    }
+
+    return r;
+}
 
 bool DirectoryCreator::directoryExists(const std::string& path)
 {
@@ -21,9 +85,10 @@ bool DirectoryCreator::createDirectory(const std::string& path)
 
 bool DirectoryCreator::removeDirectory(const std::string& path, bool force)
 {
-    return rmdir(path.c_str()) != -1 ? true : false;
-	
-	// implement force delete here too
+    if(!force)
+        return rmdir(path.c_str()) != -1 ? true : false;
+    else
+        return !remove_directory_force(path.c_str());
 }
 
 #elif defined(__WIN32__)
